Handle cases where colon is in the first path segment (#130)
`Request.url.path` is supposed to be relative to `Request.handlerPath`,
this invariant is enforced by throwing `ArgumentError` in the
constructor. However, if an incoming request has colon in the first
path segment the `shelf_io` adapter will create a request with `/`
handler path creating a `Request.url.path` that is relative to `/`
and which has colon in the first segment. This causes unnecessary
encoding of the colon which breaks the invariant check.

In hindsight it's probably not ideal to use `Uri` for holding a
relative path as this will always mangle the colon. But fixing this
is very breaking.