Ignore protocol header if handler did not specify allowed protocols (#28)
Closes #27

The spec says "if the server does not agree to any of the client's
requested subprotocols, the only acceptable value is null". The mozilla
docs for implementing a web socket server say "The client may close the
connection if it doesn't get the subprotocol it wants."

From this I take it that the safest thing to do is to ignore the
protocols. A server which wants to be more careful can handle the `null`
protocol in whatever way makes sense.

- Add a null check before searching for overlap in protocol.
- Remove the explicit check for `protocols` with 2 argument callback.
- Rename `protocols` to `requestProtocols`.
- Add tests demonstrating new behavior, remove test for ArgumentError.