Remove content-length for notModified responses (#183)
It is an error to include a non-zero content-length when there is no
body. In some places, especially those where we are implementing a
proxy, we were previously relying on behavior in shelf which would zero
out the content length when there is no body. After allowing a content
length without a body for `HEAD` requests we lost the safety for other
request types.

The errors that have surfaced since published are related to
`Response.notModified` which should never have a `content-length`, so
handling it there should retain most of the safety we had before. There
is still the possibility for problems in other response types which will
need to be handled at the usage site.