Use HttpClient header defaulting for x-powered-by
Fixes #270

The `HttpClient` from `dart:io` supports adding default header values to
requests that don't already have them. Using this mechanism for the
`x-powered-by` header retains current behavior, but allows the caller to
override it by mutating the returns server's `defaultResponseHeaders`.
If a request manually includes a header value it will be retained as
before.